refactor(client-details): useDifferentArchitecture4Component
