+++ src/components/organisms/Customers/WalletTab/wallettab.scss
@@ -37,7 +37,7 @@
         display: flex;
         align-items: center !important;
         padding: 0 !important;
-        margin: 1.5rem 0;
+        margin-top: 1.5rem;
         margin-left: -0.5rem;
 
         &-text {
