+++ src/modules/clients/containers/client-details/client-details.module.scss
