+++ src/modules/clients/containers/client-details/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./client-details";
